requirements.txt
# AI SDK
google-generativeai

# Frontend (UI)
streamlit

# Vector Database (RAG)
chromadb

# Data Processing
pandas
python-dotenv

#numpy